+++ src/commonMain/kotlin/csense/kotlin/extensions/duration/Duration.kt
@@ -1,0 +1,7 @@
+package csense.kotlin.extensions.duration
+
+import kotlin.time.*
+
+public suspend fun Duration.delay() {
+    kotlinx.coroutines.delay(this)
+}