+++ src/commonTest/kotlin/csense/kotlin/extensions/duration/DurationTest.kt
@@ -1,0 +1,21 @@
+@file:OptIn(ExperimentalCoroutinesApi::class)
+
+package csense.kotlin.extensions.duration
+
+import csense.kotlin.tests.assertions.*
+import kotlinx.coroutines.*
+import kotlinx.coroutines.test.*
+import kotlin.test.*
+import kotlin.time.*
+import kotlin.time.Duration.Companion.milliseconds
+import kotlin.time.Duration.Companion.seconds
+
+class DurationTest {
+    @Test
+    fun delay() = runTest {
+        0.seconds.delay()
+        currentTime.assert(0)
+        110.milliseconds.delay()
+        currentTime.assert(110)
+    }
+}